Ensure delta sync always covers full history window
