+++ requirements.txt
@@ -4,3 +4,4 @@ paramiko==3.4.0
 gunicorn==21.2.0
 google-cloud-tasks>=2.13.0
 google-cloud-storage>=2.16.0
+google-cloud-bigquery>=3.11.0
